test: Modify e2e tests on one-to-one relations to integration tests.
